Fix broken output logic in trap integrator
An open source user reported bizarre behavior from Xyce when he
used ".tran <tstep> <tstop> <tstart>" options in conjunction with a
".options output initial_interval=<interval>" statement.  Xyce was
outputting not only at the requested times, but repeatedly at
unrequested times.

I tracked this down to a change made as part of commit 2c83398bc from
2013, when the logic for choosing when to output was modified slightly
to support .FOUR and the fourier measure statements.  Back then, it
was necessary to continue to store interpolated solution values even
if they weren't being printed.  Unfortunately, one call to the output
manager's "tranOutput" function was overlooked and did not get the
"skipPrintLineOutput" argument passed to it, and that argument
defaulted to false if not given.  This mistake only existed in the
Trap integrator.

Adding the argument to the one function call that was missing it fixes
the problem.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1299