Add more formats for TABLE, pairs w/parens
One format I had not yet supported for TABLE was the format that has the
each time,value pair surrounded by parens, and with no comma in between
each pair.  In other words:

TABLE { V(2) * (V(1) + 30) / 60 }  (0.0000000, 0)  (0.0312500, 0) ..

The TABLE operator has a number of different formats and I don't think
I've covered all of them yet.   But this was an important one, as it
comes up early in the regression tests, for the Bsrc_C1 test.  So, it
kept failing and it was noticable.  Fixed now.

There will be others to support that I haven't done yet.  An important
one is the "tablefile" command.

I should note that at the moment, it feels like I probably have too
many  rules in the Bison parser (ExpressionParser.yxx)  for TABLE.  I
should be able to reduce the number and simplify this section, once I
spend a little more time on it.