Change phase op from "P" to "Ph", create power op
I had chosen to have "phase" be tokenized with P.  ie, P(V(1)) means
"phase of V(1)".  However, this conflicts with P for power, which is
what Xyce uses.

So, I've changed the phase token to be Ph, and I've created a place
holder operator for Power.   It doesn't do anything yet, but the parsing
will be more correct now when running Xyce test suite.