Fix slightly broken logic in output interpolation
Since time immemorial, there has been a boolean whose value is
meant to flag that the first output time of a simulation that had
"tstart" specified on a .tran line should NOT invoke output
interpolation.

This logic has always included "beginningIntegration" as one of the
things anded together, and in many cases where one might combine use
of tstart and .options output, the first output is NOT happening when
beginningIntegration is true, and should NOT trigger interpolation to
previous times.

This issue turned up when I was trying to create good tests for bug
1299, where an open source user had found that we were generating some
weird output when he used tstart and .options output together.  I
fixed one of the issues yesterday that addressed his immediate needs,
but when I went to create thorough regression tests for the fix I
found that there were some cases that weren't working properly.  This
commit fixes those, too.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1299