update ExpressionData::setup for attachFunction
I previously modified parts of the IO package and also the sensitivity
capability to "attach" external functions and global parameters, instead
of using string copies.  I had not done this modification yet to the
ExpressionData class, but it needed it.

The code in the expressionData class "setup" function resembles (at
least in the relevant parts) the code in the setupObjectiveFunctions for
.SENS.  So, this was mostly a code copy with a few name changes.  Seems
to work, but needs more testing.

There is MUCH more refactoring to be done in the ExpressionData class to
work properly with newExpression.