Duplicate OffTime Measure Files as New TranStats Class Files
This commit copies the files for the OffTime measure class to the
files for the TranStats class.  That new class will be used as a
base class for the OffTime, OnTime and Frequency classes.  This
duplication should preserve the history/blame of the TranStats
class as the history for the new TranStats class.

SON Bug 1285
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1285