Refactor FREQ, OFF_TIME and ON_TIME Measure Code
This commit refactors the code for the Frequency, OffTime and OnTime
classes to put all of their common functions into the new TranStats
class that is derived from the Base class.  It also simplifies
the Frequency, OffTime and OnTime classes so that a common function
is called (in each of those classes) by the UpdateTran() function
in the TranStats class.  This should eliminate most of the duplicate
code in those three classes.

SON Bug 1285
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1285