Creating a new group for use with .PRINT
I have gone back and forth on how to handle the "group" concept for the
newExpression library.  Most recently, I had decided to have a single
monolithic group that was used by all expressions.  I implemented
(incomplete, preliminary) as the "mainXyceExpressionGroup".

I have now decided that having a single monolithic group for everyone
was not the right choice.  The expressions used in .PRINT have different
requirements than the expressions used in device parameters and Bsrc's.

.PRINT line (as well as other outputs like .measure and .sampling)
outputs have the following requirements:
 - they don't need derivatives.

 - they aren't needed during initial setup.  So, certain
    convenient data structures that come from the output
    manager can actually be epxected to work.

 - they output lots of things that could (probably) never
    be used in a device parameter or Bsrc, such as noise, lead currents,
    etc.

 - UQ operators like Gauss and Agauss probably
   aren't needed.  Those operators are typically
   applied to inputs, not ouputs.  As such, this probably
   doesn't need to be a singleton.  Instead can be
   a group allocated by each ExpressionData object, etc.
   That will make some of the issues for it a bit simpler,
   probably.

 - they only need to work on proc 0.

So, I am now attempting to have 2 main group classes.  (there may be
reasons for others, later).  These 2 are now:

mainXyceExpressionGroup - used for device parameters: needs derivs and
support for sampling, must work on all processors.  Does not need
things like dno, onoise, etc.

outputsXyceExpressionGroup  - used for .PRINT and other outputs : no
derivatives, no sampling, but needs lots of derived outputs.  This group
will mostly be used from the ExpressionData class, and will probably use
as much of the Op classes as possible.

There are some things that have only been previously implemented for
.PRINT, that I want to have work with parameters and devices. One of
the most important of these things is the use of complex numbers.  A lot
of these types of outputs were described with the old expression library
as NODAL_COMPUTATION.    Much of the machinery for these types of
operators can now just be handled in the new expression  library.    So,
this will require some re-thinking.