Start process to replace get_names function calls
The get_names function is used a LOT with the old expression library, in
a lot of overloaded way.

With the new expression "group" model, this function isn't as neccessary
as before.  *Some* of the functions it provides are still necessary,
particularly for resolving an expression's external dependencies, such
as with .func, .param, or .global_param.  I have chosen to *not* have those
resolutions handled by the group, so some accessors related to them are
still needed.

so, my goal is to ultimately get rid of all the "get_names" function
calls, and then get rid of the function altogether.  In this commit, I
have removed all the remaining get_names calls from the sensitivity
class.