Lots of cleanup, better support for specials.
I started doing a lot of code cleanup and of course while doing it
noticed something I needed to fix. (which was part of the point of
cleanup)

As I did cleanup, I noticed a comment indicating that specials (like
TIME, TEMP, FREQ, VT) had not yet been handled for attached nodes.  In
other words, for expressions with external dependencies, the new
expression code didn't properly report if a special was being used or
not.   This was one of those things I made a note of and meant to get to
later.  Well, now is that later.

This commit goes part of the way towards fixing it, but not quite all
the way.