Fix for external dep of specials(time,temp,etc)
In my recent cleanups I discovered that I neglected the use case of
external dependencies that included specials like time, temp, vt and
freq.   A recent commit took one incomplete step in the direction
of fixing this issue.

This commit fixes it for real.  The previous commit only considered
getting the time/temp/vt/freq dep booleans set properly.  But it
neglected the fact that each expression object would have its own unique
allocation of a special.  Ie, the specials are singletons inside of an
expression, but they are not singletons from one expression to the next.

In this commit, I do a traveral of the expression tree and gather all
the specials.  Then when I grab specials values from the group, I apply
them to all the gathered objects.  This is the first time I've set up a
"expression" traversal rather than a AST traversal.  Going foward, I may
try to eliminate all the AST traverals (to the extent possible) and
replace with expression traversals as I think in most cases they will be
cheaper to execute and easier to maintain in the source code.