Initial attempt at optimizing for constant expr
If an expression does not depend on anything then it should not need to
be re-evaluated.  The old expression library had many checks in it to
ensure that it only re-evaluated when it had to.  I didn't initially
bother with this, and for some regression tests this matters a lot.

This commit is a first pass at adding this kind of logic to the new
expression library.  It helps, but it isn't enough.