Refactor printMeasureResult() Functions
This commit refactors the printMeasureResult() functions to
remove the Boolean flag in their function definitions.  The
printMeasureResult() functions now handle the output to
the mt0, ma0 and ms0 files, while the printVerboseMeasureResult()
functions now handle the more descriptive output to stdout.
This approach is a bit cleaner, and makes it clearer what the
functions actually do.

SON Bug 1285
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1285