Add further optimization for constant expressions
In a previous commit I added some code to the newExpression class, to
check if an expression is "constant" or not.  If it is constant, then it
doesn't need to be re-evaluated after the first evaluation.

I was using the LUMPEDLINE/baselineLumpedLine.cir circuit as a baseline.
I used this netlist as I noticed that the LUMPEDLINE regression test was
running really slowly with the new expression library and I started
digging to find out why.  The main reason was that the function
"evaluateFunction" in the new expression library was being called a lot,
unconditionally.

So, the above optimization from my prior commit helped.  Prior to
the above optimization the new expression library took 10's of seconds
on this test (as it has many, many constant expressions), and the old
expression library took (roughly) 3 seconds.   After that optimization,
the new expression library took about 14 seconds.

In this commit, I've added further optimzation to the function in the
device entity class "updateDependentParameters".  There a several
versions of this function, and I've only modified one of them.  In the
function I modified, there is some logic at the top to determine the
"changed" boolean.  At an earlier stage of my work I punted on setting
up this boolean, and just set it to be unconditionally true.

Now, I was able to go back and handle it somewhat more correctly.  Now
it will query each expression parameter "getIsConstant" function.  If it
is considered constant, then the changed boolean is set to false, and
the evaluateFunction function isn't even called at all.

With this optimization, the runtime of LUMPEDLINE/baselineLumpedLine.cir
got down to around 7 seconds.  Still not good enough, but a real
improvement, and much closer to the ballpark.  Possibly the optimization
flags used for my two builds are not the same, so I should check that.
That seems to be the most probably explanation at this point for the
difference of 3 vs 7 seconds, given that the new expression library is
barely used now in the calculation.

While it is a little hard to tell (as there are many, many broken
nightly regression tests with the new expression library, mostly having
to do with expressions on the .print line), this optimization doesn't
appear to have caused any new broken regression tests.  And, the test
suite runs noticably faster.