Update ast unit tests for new spice sources
Recently I changed the spice sources so that they could have a variable
number of arguments.  This change required that most of the arguments
be passed into the spice source AST node constructors as a vector of
nodes, rather than having each source arg be a separate function
argument.

When I did this, I was mostly focused on the parser level unit tests.
So, I didn't update the AST unit tests, which all broke due to this
change.  That is fixed now.