Better fix for ground nodes.
This fixes the onebit.cir test, as well as (probably) others.

The old expression library automatically detects ground nodes during
parsing and sets the to zero.  So, when the calling code requests a list
of names from the old expression library, any ground nodes are excluded
from the list.

In the new expression library, I didn't initially do that.  So, in
regression test that used expressions which depended on ground, there
was potential for problems, particularly if they were ground nodes
inside of subcircuits.  During setup of Bsrc's and other
expression-based devices, any extDepVars were being looked up in the
topology directory.  A node such as X5:X3:0 would not be found in that
directory, resulting in a fatal error.

This commit enables the new expression parser to trap for ground nodes
when parsing V() operators.  If the node is NOT ground, then a voltage
op is allocated as before.  But if it is ground, then a numvalOp is
allocated instead, with a value of zero, and it is not added to the
voltageOpVec container.  As a result, it is never returned via
get_names, but it is also not included as one of the values requested by
the group during the evaluate phase.