Refactor Duty Measure Class to use TranStats Class
This commit refactors the Duty measure class to use the TransStats
class, like the Frequency, OffTime and OnTime classes which are also
variants of duty-cycle calculations.  A side-effect of this refactor
is that the DUTY measure no longer supports the RISE, FALL and CROSS
qualifiers.  This change is okay, since those three qualifiers really
aren't very useful for the Duty, Frequency, OffTime and OnTime
measures.  (Note: the RISE, FALL and CROSS qualifiers were never
implemented for the Frequency, OffTime and OnTime classes.  So, this
commit makes the behavior of all four measure types consistent.)

SON Bug 1285
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1285