Add support for "gmin" special.
In the original design of new expression, I set up for "specials".  They
were TIME, TEMP, VT, and FREQ.

As it turns out, I should have also included GMIN.  This commit adds it.
This commit also includes some cleanup, including removal of some old
gmin code from the circuit context, as well as a bunch of other old
obsolete expression functions.

For context:

The work on GMIN was inspired by my attempts to get rid of some of the
old expression library functions.  One of them was a "replace_var"
function, which took an "op" as the second argument.  I didn't want to
make the new expression library understand Ops directly, so I wanted to
remove the function.  It turns out the only place in Xyce that called
this function was in the parser, in one place, where the "gmin" and "vt"
ops were parsed and processed in expressions.

That call isn't necessary anymore, so the funciton is gone.  I have also
removed a number of other functions that aren't needed any more.