Light Re-Factoring of DerivativeEvaluation Class
This commit does a light re-factoring of the DerivativeEvaluation
class to move some duplicated code into some common functions
that can be shared by the updateTran(), updateDC() and updateAC()
functions.  The bigger re-factor of the handling of the AT and
WHEN conditions was not done yet.  That handling is the same within
updateDC() and updateAC().  However, the handling in updateTran()
is slightly different because RISE, FALL and CROSS are supported
for TRAN mode measures.  The final change is that the doneIfFound
local variable was redundant with the measureLastRFC_ private
variable; so it was removed.

SON Bug 1285
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1285