Fix for new expression with STEP and SAMPLING
The main thing this commit accomplishes is to implement a "getVariables"
function and call it from a few places.

The "get_names(XEXP_VARIABLES, variables)" funciton calls were not doing
anything, and this was causing a bug.  The bug was that .STEP didn't
update global parameters properly.  Same with .SAMPLING.  So,  many
tests related to these features failed.

The reason that getVariables fixes it is this; in the IO package, there
are if-statements to detect if an expression valued parameter has
various dependencies, including w.r.t. what it calls "variables".
"Variables" in the old expression library mean global_parameters,
basically.

getVaribles is not implemented, and will push the names of parameters
onto a vector of strings.  The parameters it will push_back onto this
vector must have a "true" value for the "isVariable" boolean.  Other
params (like ones set to const) will not be put on this vector.