Refinement to new getVariables function.
I previously had this function only return names of parmaeters that had
been designated with "isVariables=true".  This wasn't good enough and
here is why:

What the function needs to do is return a list of param names used by
this expression that have external dependencies and are not considered
constant.  If they are constant, they are probably .params

This getVariables function (and the get_names(XEXP_VARIABLES) before it)
is mostly called on parameters that are used in device instances and or
models, when Xyce is trying to decide if a model parameter should be
considered a "dependent param".  This logic happens in Xyce after all the
.param and .global_params have been set up, and all the parameters have
been resolved in this expression.  If it happened earlier, it would
possibly give the wrong answer.

Params in the new expression library have 3 possible states:
isConstant, isVariable, and isAttached.

"Constant" means that it is probably a .param, and is set during
the parsing process, and will never change.  So any variable with
this boolean set to true will not be returned by this function.

Variable" means that the parameter can be set to a simple number,
but that this simple number can be changed during the simulation.
So, this is a .global_param, but it is a simple one, that is just
a number.

"Attached" is the biggest departure from the old expression library,
and it means that the parameter object in this expression points
to the top node of another expression class.  A parameter with this
status technically could, if you follow the attached AST all the
way down, still be "constant", if the attached tree is constant.
But it could also be variable, and as of this writing I haven't
set up an easy way to check this.  So, I just assume for now that
attached nodes are variable, unconditionally.  I should fix this
later, as this is overkill.

In this function, "Variable" and "Attached" parameters have their
names added to the std::vector of strings.  This probably doesn't
100% produce the same result as the old expression library, but
it is close enough for now.

With this fix, a few more .STEP and .SAMPLING (and other tests that rely
on changing .global_params) now pass.

There are many, many tests that still fail b/c the expressions on the
.print line are currently intentionally broken.