Update to handle another table specification
The "power_e_gear.cir" regression test was failing.  One reason it was
failing was that the new expression library could not parse this:

{TABLE {V(1,0)} = ( 0 , 1 ) ( 1 , 2 )}

It could, however, parse this:

{TABLE {V(1,0)} ( 0 , 1 ) ( 1 , 2 )}

So, I needed to add another parser rule to handle the first case, which
is what is in power_e_gear.cir.  That is done, as of this commit.

I have also added a new parser unit test, to test this.  It is this one:
TEST ( Double_Parser_table_Test, power_e_gear)

One thing I noticed, but didn't fix, is that the interpolators  used by
the tableOp attempt to extrapolate when given a value outside of the
original domain of the the specified table.  In the old expression
library, it doesn't do this.

So, for example, in the above expression, if the input is between 0 and
1, then the output will be between 1 and 2.  However, if the input is
outside of 0 and 1, then in the new expression  library it returns a
number outside of 1 and 2.    In the old  one, it will just return one
of the bounds. (1 or 2)