Checkpoint commit; fixed several things
This commit is larger than I originally intended so it includes a bunch
of changes.  It also includes a bunch of debug print statements that I
need to get rid of later.  It includes several days of work....

The first thing that this includes is yet another fix for "getVariables"
function in the expression library.    This was intended to replace
"get_name(XEXP_VARIABLES,names)".  It has been harder to get this right
than I originally expected.  It is basically supposed to return a list
of names  (std::vector of std::strings) of the parameters in this device
that are exernal dependencies and subject to change.   (i.e. global
params).  I eventually decided that the only way to do this was to
simply have a boolean in the paramOp class indicating if it was a .param
or not.  If it is a .param, then it is excluded from what is returned by
getVariables.

Another change is that I removed many, many unused (or nearly unused)
functions from N_UTL_Expression.  This also required that I change some
parts of Xyce to call the new replacement functions (such as
getVaribles, etc).

Finally, this includes the start of *real* code development in the
outputsXyceExpressionGroup class.  Up until this point it was just a
copy of the mainXyceExpressionGroup.    it is not anymore, and is mostly
attempting to reproduce the "Op" based machinery of the ExpressionData
class.  This still needs a lot of work, but it is getting there.