Make B4SOI SHMOD parameter instance/model
The documentation for the SHMOD parameter in the BSIM SOI 4.6.1 manual
says it should be both instance and model, but the Verilog-A specified
it as instance (there is no way to specify "both" in standard
Verilog-A).  I neglected to add the Xyce/ADMS hackage necessary to make
it both instance and model.

https://gitlab.sandia.gov/xdm/data-model/issues/207
xdm/data-model@#207