mucho development on leads,power, noise for .print
A lot of work happened today.

The new group, outputsXyceExpressionGroup has been substantially
improved.  It supports .print line outputs.

The outputs of noise, power and lead currents needed a lot of work
before today.  They still probably need lots of work, but the work today
represents a substantial improvement.

As an example, this commit correctly implements most of the noise
operators, and now most noise tests with expressions will pass.  Note, I
had not realized that DNO and DNI sometimes take two arguments, and now
this feature works.

Lead currents had a lot of problems.  The logic for differentiating
between lead and Vsrc currents was busted.  This is mostly fixed but
probably needs a bit more work.  The code for obtaining lead currents
from the group was busted as well, in  multiple ways.  Mostly fixed now.

Also, lots and lots of code cleanup.