Fix Handling of Failed 1-Point AVG and RMS Measures
This commit fixes the handling of AVG and RMS measures that have
only one valid point in their measurement windows.  Those measures
are "failed" by definition.  The previous fix for this issue made
the bone-headed mistake of modifying the initialized_ private
variable in a "getter" function.  Naturally that caused a "side
effect", in that then using an RMS or AVG measure on a .PRINT line
caused that measure to always "fail".  This commit uses a better
approach, wherein the new numPointsFound_ private variable tracks
how many points were found in the measurement window.  This allows
the AVG and RMS measures to do the correct thing.  (Note: the INTEG
measure should, and does, report 0 for a 1PT measure).

SON Bugs 1267 and 1282
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1267
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1282