Fixed a bunch of valgrind errors, from recent work
Fixed about 10 different things that came up during the past week or so.

Most of them were caused by not deleting memory that needed to be deleted.
Frequently in the parser file (ExpressionParser.yxx) I forget to do this,
since I'm mostly thinking in terms of refCountPtr's.  However, since I'm
using *pointers* to refCountPtr's, those pointers do have to be deleted.

Another small mistake: In the way I designed the class, there is one
"master" vector of pointers-to-RCP of AST nodes that has to be cleared
and deleted in the destructor.  All other vectors of AST nodes are
vectors of RCPs themselves (rather than pointers to them) so these
do not have to be deleted, ever.

I had one vector that *should* have been a vector of RCPs, but instead
was a vector of *RCP.  That vector was the srcAstNodeVec_ object.
Valgrind was complaining that this vector was not being deleted, in cases
where it was copy constructed.  But in cases where it was NOT copy
constructed (or assigned), it should not have been.  So, deleting it
correctly was dependent on if it had been copy constructed or not.
This was exactly the kind of headache I had hoped to avoid by using
RCP.  Fortunately, as I noted, there was no reason for this vector
to be a vector of *RCP rather than RCP.  So, I have changed it to a vector
of RCP and it is all good now.  No delete is needed, and the valgrind
error is gone.

All clear now.