Various fixes for tableOp class
This commit implements two things.

(1) I had previously not set up the tableOp to return breakpoints, and
now it does.  It only returns them if the input is "time".   It does a
somewhat brain dead thing and returns a breakpoint for everything time
point in the specified table.  This is what the old expression library
does, approximately.

(2) The tableOp was doing the wrong thing if the input was outside of
the bounds of the table.  The way it was written, it would continue to
extrapolate.  This was wrong, and caused many things to break.  It now,
correctly, just uses the table value from the edge of the table, and
does not attempt to extrapolate.