Refine support for solution var currents
Currently the new expression library kind of treats solution currents
and lead currents separately.  This probably isn't the best organization
in retrospect, and I may combine them later to simplify things.

But for now they are separate, and this matters more for the parameter
(and thus Bsrc) usage of current variables in expressions than it does
for .print usage of current vars in expressions.  I had initially set up
the parser to only recognize Vsrc currents as valid solution currents.
This commit expands it to things like e-sources and h-sources, etc.