Fix params to change from expr to values
Many of the remaining tests that were failing in the test suite were for
tests that complained about a parameter of a non-double type begin of
type EXPR.  ie, parameters can only be EXPRession parameters if their
fundamental type is double.

but, there were many examples in the test suite of non-double parameters
being set with simple expressions.  GRORDER (an int) in the photocurrent
models is a prominent example.   It was common in many tests to see
things like:

.param grorder=27
grorder={grorder}

The second statement would technically have an expression on the RHS,
but of course, a trivial one.

The reason that the old expression library wasn't having this problem
was actually just due to an efficiency in the circuit context class that
I inadvertantly removed.    In the "resolveParameter" function, if a
parameter is completely resolved, and doesn't have any function
arguments (i.e., isn't actually a .func), then the parameter is
de-expressionized, and just set to its value.

As long as this (the de-expressionization) happens, then the downstream
errors (of non-double type) don't happen either.

My recent work changed so many things in the circuit context that I
overlooked this issue and broke it.   Part of the reason was that I
split the resolveParameter function into 2 functions. One that was for
.funcs ONLY, and another for everything else.  But, I unfortunately left
the "de-expressionization" code in the .func-ONLY version of the
function, where it did not good, and deleted it from the "everything
else" version of resolveParameter.

This is fixed now.  The de-expressionization has been restored to the
correct place in the circuit context and many more tests pass now.