Support for Hspice version of "limit"
I tried to account for various Hspice operators in the expression
library, but overlooked this one.  Like with the old expression library,
I implemented a 3-argument version of limit first.  But Hspice has a
2-argument version, and it refers to a distribution function.

In the old expression library the 2-arg limit is implemented as
limit(x,y)=x+y, so for now I have done this as well.