switch to default=complex, fix temp. handling
This commit includes several changes:

(1) switch from default data type from double to std::complex<double>

The AST classes are all templated, and can be compiled with either
double or std::complex<double>   I have not tried any other data types.
I'm sure some could work, but not relevant at the moment.

I initially had a fantasy of having the invocation of the expression
library be a run-time decision, dependingn on the use case.  So, if
running a DC sweep, you probably would not need a std::complex typed
expression tree.   However, it turned out that I couldn't (easily,
anyway) do this.  Flex and Bison do not seem to provide a method for
producing templated functions and/or classes.  Possibly as I learn more
about flex and bison, I will discover a way to do this, but for the
moment I don't see it.

So, as I am (for now) using flex/bison to handle parsing, I have to know
the data type at compile time, not run time.  So, either all expressions
are of double type, or all expressions are of complex type.

As complex will be needed some of the time, I've chosen to have
std::complex be the type used in the expression tree.  However, this
decision is something I arrived at only recently.  In my earlier work, I
had made the default type to be double.  This was set in the file
"ExpressionType.h", and the typedef is "usedType", which you can see
throughout the expression source files.    If I wanted to compile the
expression library using complex, then I had to set a pre-processor
macro on the compile line:  -DUSE_TYPE_COMPLEX.   For all of my recent
work I have used this.     I haven't built Xyce with a "double" typed
expression library in several weeks.

I finally decided to change the default.  Now, when the expression
library is built, it will use typedef usedType std::complex<double>
unless you tell it otherwise.  If you want to build with "double", you
now must use -DUSE_TYPE_DOUBLE.    I made this possible to support the
unit test program "parserUnitTest.C", which uses double.

(2) Disable the conditional evaluation of an expression based on the
"isConstant_" boolean.   A little while ago I tried to make the library
more efficient, and created this boolean.  If an expression was deemed
"constant" this boolean was true, and that would mean the expression was
only evaluated once.  Unfortunately, my logic for setting up
"isConstant_" boolean was flawed, and so it sometimes resulted in
expressions NOT being evalated when they needed to be.  So, until I sort
out what is wrong with that, I have disabled this conditional.

(3) fixing the use of TEMP in several ways.  This was driven by failures
observed in the BUG_1203_SON tests.  The main problem was how CONSTCtoK
was applied inside the expression.  This happened in the device entity
under the following conditions:

  - the expression is the RHS of a device "temp" parameter
  - when the units need to be "auto converted" from C to K units.

For the old expression library, this was accomplished by
getting the string of the original expression and then modifying it to
include "+ CONSTCtoK", and then using that new string to create a
new expression.

This approach doesn't work with the new expression library, b/c the
new expression library doesn't handle external dependencies via
string replacements.

The CONSTCtoK modification to the string was happening in the device
entity, long after the IO package was done with its work on parameters.
And, thus, long after and function resolutions.

With the new expression library, doing a string modification for
CONSTCtoK could not include those IO-based param and func resolutions.
They were lost, because they were not in the string.

The right way to fix this is not by re-parsing with a new string.
Instead, it is better to just modify the AST, which is what Xyce
now does with this commit.