Add another table format and test
This is for TABLE(x,1,1,2,2,)   <- where there is a comma after the last
value, but b4 the right paren.   My parser rules didn't allow that, but
it is used in at least one regression test.

One could argue that the above format should not be allowed, as the
final comma implies there should be another value.  But, apparently the
old expression library allows it, so there it is.