initial work on autodetecting breakpoints
The old expression library has a breakpoint detection algorithms,
contained in functions such as "get_breaks" and "get_break_time_i".
(Horribly, misleadingly named functions, BTW)

get_breaks returns a Newton iterate for whatever nodes exist in the
current expression tree.  It does NOT return a breakpoint, as implied by
the function name.  The function "get_break_point_time_i" uses these
Newton iterates as part of its procedure to auto-detect breakpoints
(discontinuities).

The problem with the old library was that it used this procedure for
everything; even for things like "spice_pulse", where the breakpoints
could easily be known ahead of time, and were not a function of the
state of the system.

So, in the new expression library I have set things up differently.  For
functions like "spice_pulse", where all breakpoints can be known ahead
of time, I do not run an auto-detect algorithm.  I just compute
breakpoints in the same manner as the N_DEV_Vsrc/N_DEV_ISRC.

However, there are some AST nodes for breakpoints are a function of the
state of the system.  The simplest example is the STP(x) function, which
returns a 0 if x<=0, and a 1 if x>0.    In the stp.cir regression test,
stp is invoked in a Bsrc as stp(time-delay), but of course many more
complicated arguments could be given to it, such as stp(v(2)), where
v(2) is a voltage node that varries with time.

With this commit, I have modified the code a little bit to add a
"getBreakPoints" function to the stpOp class.   I have also added a
little bit of logic to newExpression::getBreakPoints, to account for
this.   The stpOp::getBreakPoints function doesn't do anything yet.  I
have been staring at the old expression library, and also reading
academic papers on event detection, to try to figure out the best path
forward on this.