Remove superfluous verbiage from fixed defects table
The table produced by fixedDefectsTbl.tex is captioned "Fixed
Defects", and is in a section of the release notes entitled "Defects
Fixed in this Release."

For years it has been a habit shared by most members of the team ---
including me --- to add the unnecessary text "This issue has been fixed."
or something like it to the description of bugs in this table.  The
presence of the bug in the table means it's been fixed, and so we
should not have to say so again.

For the last few releases I've been trying to refrain from adding this
text to things in this table, and have been removing these words when
they show up.  Let's get out of the habit.  All that need be present
in this table is a description of what was fixed.