fix for thermal resistor with newExpression
Some of the regression tests that use the thermal resistor device were
failing with the new expression library. It took me a while to figure
out why.  I did most of my debugging on the THERMAL_RESISTOR/linear.cir
test, but my fix affected a few other tests, such as
BUG_1826/linear_simple.

The issue in the linear.cir circuit had to do with the fact that the
linear.cir test relied on a model card, which made the parameters
"resistivity" and "heatcapacity" expressions which were dependent upon
temperature.  They used the table capability, as well, but I think the
breakage that I fixed would have applied to most temp-dependent
expressions.  There were two problems:

(1) The main problem was that when newExpression evaluated these
expressions, the temperature it was using was NOT the local temperature
of the thermal resistor, but rather the more global "circuit"
temperature that it obtained from the device options class.  By default,
when someone uses "temp" in Xyce, this is almost always the temp that is
intended.

The way that the resistivity and heatcapacity received the local
temperature was via the "updateDependentParameters" function,
specifically the version that takes "double tempIn" as a function
argument.  The thermal resistor is the ONLY device model in Xyce that
invokes this function, BTW, which is part of why I didn't fix it
correctly until now.

To get this function to play nicely with the new expression library, I
had to restore a "set_temp" function to it.  I gave it a slightly
different name "setTemperature" to differentiate it.  Most values that
used by the new expression library from Xyce come from the "group"
class.  Same is usually true of temperature.  But there is no easy way
for the group to know which temperature is intended.  So, the easy way
to fix it was to have this one instance where temperature is "pushed"
into the expression, rather than pulled by it via the group.  If this
function is called, then the "group" version of temperature is not used.

This fixed part of the problem, but not all of it.

(2) the issue of "changed" and when/if to re-evaluate the expression.
The old expression library has a lot of logic to determine if its
dependent values have changed since the last evaluation.  If they
haven't changed, then it isn't re-evaluated.  I have started to
implement something like this in the new expression library, but the
logic it uses is somewhat broken, so I have disabled it for now.
I had thought that this was only an efficiency issue, but it turns out
that it isn't, at least in this case.

The DCOP calculation is unable to converge if the
resistivity and heatcapacity are re-evaluated at every Newton iteration.
They affect the internal temperature, so you get a feedback loop that
never ends.

It doesn't quite make sense, but the logic of the "changed" tests in the
old expression library (in conjunction with code in the device entity)
determined that the temperature wasn't changing during the DCOP
calculation, and thus these 2 parameters are constant during the DCOP
phase.   How they determined this, I never quite figured out.  But
determine it they did, and this allowed this very simple circuit to
converge.

In the new expression library, they continued to be re-evaluated, and
convergence wasn't achieved.  I attempted to add some "changed" logic
with respect to setTemperature, but that didn't fix it.

Instead, I chose to conditionally call "updateDependentParameters(temp)"
from the thermal resistor model if the device was NOT in the TRANOP
state (ie, if the analysis is transient, and we are in the operating
point initialization phase).  Once I did this, all the thermal
resistor tests pass.

This conditional will not apply to DC sweeps or OP calculations.  But
none of the current regression tests appear to use the thermal
resistor in a purely steady state calculation.